requirements.txt
uvicorn
openai
fastapi==0.111.0
gunicorn==21.2.0
langchain-community==0.2.5
langchain-experimental==0.0.49
langchain-openai==0.0.6
pip-chill==1.0.3
python-dotenv==1.0.0
chromadb
pypdf==4.2.0
pysqlite3-binary
sentence-transformers==3.0.0
langchain-huggingface